Made the game work
Added end turn button, moves remaining text, units move in a straight
line, when a king dies the other player wins, tiles highlight correctly,
notes are noted, listeners are listening, it practically is a game at
this point.

+++ src/Entity.java
@@ -3,7 +3,7 @@ import java.awt.Graphics2D;
 
 public abstract class Entity {
 
-	protected int x,y;
+	protected double x,y;
 	
 	abstract void render(Graphics2D g);
 }
